# swagger-ui/api-docs.yaml
openapi: 3.0.3
info:
  title: Todo App REST API
  description: |
    A comprehensive REST API for a Todo application built with Express.js and JSON Server.
    
    ## Features
    - User registration and authentication
    - Todo CRUD operations
    - Category management
    - User-Todo relationships
    - JWT-based authentication
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@todoapp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api-proxy
    description: Proxied API Server (Recommended for Codespaces)
  - url: http://localhost:5000
    description: Local Development Express Server
  - url: http://localhost:3001
    description: Local Development JSON Server
  - url: http://localhost:3001
    description: JSON Server (Database)
...
paths:
...
  /api/todos:
    get:
      tags:
        - Todos
      summary: Get todos
      description: Get todos for the authenticated user with filtering and pagination
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by todo status
          schema:
            type: string
            enum: [pending, in-progress, completed]
        - name: priority
          in: query
          description: Filter by todo priority
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: category
          in: query
          description: Filter by category name
          schema:
            type: string
        - name: search
          in: query
          description: Search in title, description, and tags
          schema:
            type: string
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Todos retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      todos:
                        type: array
                        items:
                          $ref: '#/components/schemas/Todo'
                      pagination:
                        $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Todos
      summary: Create todo
      description: Create a new todo for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoInput'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Todo created successfully"
                  data:
                    type: object
                    properties:
                      todo:
                        $ref: '#/components/schemas/Todo'
  /api/todos/{id}:
    get:
      tags:
        - Todos
      summary: Get todo by ID
      description: Get a specific todo by its ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
      responses:
        '200':
          description: Todo retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      todo:
                        $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Todos
      summary: Update todo
      description: Update a todo completely
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoInput'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Todo updated successfully"
                  data:
                    type: object
                    properties:
                      todo:
                        $ref: '#/components/schemas/Todo'

    delete:
      tags:
        - Todos
      summary: Delete todo
      description: Delete a todo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
      responses:
        '200':
          description: Todo deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/todos/{id}/status:
    patch:
      tags:
        - Todos
      summary: Update todo status
      description: Update only the status of a todo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [pending, in-progress, completed]
                  example: "completed"
      responses:
        '200':
          description: Todo status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Todo status updated successfully"
                  data:
                    type: object
                    properties:
                      todo:
                        $ref: '#/components/schemas/Todo'
  /api/todos/bulk-delete:
    post:
      tags:
        - Todos
      summary: Bulk delete todos
      description: Delete multiple todos at once
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - todoIds
              properties:
                todoIds:
                  type: array
                  items:
                    type: string
                  example: ["1", "2", "3"]
      responses:
        '200':
          description: Bulk delete completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Successfully deleted 2 todos"
                  data:
                    type: object
                    properties:
                      deleted:
                        type: array
                        items:
                          type: string
                      errors:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            message:
                              type: string

  # Category Endpoints
...
  /todos:
    get:
      tags:
        - Relations
      summary: Get all todos (JSON Server)
      description: Direct access to JSON Server todos endpoint
      parameters:
        - name: userId
          in: query
          description: Filter todos by user ID
          schema:
            type: string
      responses:
        '200':
          description: Todos retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Todo:
      type: object
      properties:
        id:
          type: string
          example: "1"
        userId:
          type: string
          example: "1"
        title:
          type: string
          example: "Complete project documentation"
        description:
          type: string
          example: "Write comprehensive API documentation for the todo app"
        status:
          type: string
          enum: [pending, in-progress, completed]
          example: "in-progress"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          example: "high"
        category:
          type: string
          example: "Work"
        tags:
          type: array
          items:
            type: string
          example: ["documentation", "api", "urgent"]
        dueDate:
          type: string
          format: date-time
          nullable: true
        reminderDate:
          type: string
          format: date-time
          nullable: true
        estimatedTime:
          type: integer
          nullable: true
          description: "Estimated time in minutes"
        actualTime:
          type: integer
          nullable: true
          description: "Actual time spent in minutes"
        isRecurring:
          type: boolean
          example: false
        recurringPattern:
          type: object
          nullable: true
        parentTodoId:
          type: string
          nullable: true
        attachments:
          type: array
          items:
            type: object
        comments:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              text:
                type: string
              createdAt:
                type: string
                format: date-time
        dependencies:
          type: object
          properties:
            blockedBy:
              type: array
              items:
                type: string
            blocking:
              type: array
              items:
                type: string
        completedAt:
          type: string
          format: date-time
          nullable: true
        isArchived:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TodoInput:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Complete project documentation"
        description:
          type: string
          maxLength: 1000
          example: "Write comprehensive API documentation for the todo app"
        status:
          type: string
          enum: [pending, in-progress, completed]
          example: "pending"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          example: "medium"
        category:
          type: string
          maxLength: 50
          example: "Work"
        tags:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 30
        dueDate:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 5
        totalItems:
          type: integer
          example: 47
        limit:
          type: integer
          example: 10
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: "Validation failed"
            details:
              type: array
              items:
                type: object